# apps/backend/static/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "summary": "Search / browse users",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "cursorSortValue",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursorCreatedAt",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "created_at",
                "follower_count"
              ]
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateJoinedStart",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "dateJoinedEnd",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "avatarKey": {
                            "type": "string"
                          },
                          "bio": {
                            "type": "string"
                          },
                          "followerCount": {
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "username",
                          "displayName",
                          "avatarKey",
                          "bio",
                          "followerCount",
                          "createdAt"
                        ]
                      }
                    },
                    "cursorSortValue": {
                      "type": "string"
                    },
                    "cursorCreatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "cursorSortValue",
                    "cursorCreatedAt",
                    "hasMore"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "summary": "Get a user's profile",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "getUserById",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    },
                    "emailId": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "bio": {
                      "type": "string",
                      "nullable": true
                    },
                    "avatarKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "bannerKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "followerCount": {
                      "type": "number"
                    },
                    "followingCount": {
                      "type": "number"
                    },
                    "postsCount": {
                      "type": "number"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "emailId",
                    "displayName",
                    "bio",
                    "avatarKey",
                    "bannerKey",
                    "followerCount",
                    "followingCount",
                    "postsCount",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/me": {
      "get": {
        "summary": "Get the authenticated user's own profile",
        "tags": [
          "User"
        ],
        "parameters": [],
        "operationId": "getMe",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    },
                    "emailId": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "bio": {
                      "type": "string",
                      "nullable": true
                    },
                    "avatarKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "bannerKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "followerCount": {
                      "type": "number"
                    },
                    "followingCount": {
                      "type": "number"
                    },
                    "postsCount": {
                      "type": "number"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "emailId",
                    "displayName",
                    "bio",
                    "avatarKey",
                    "bannerKey",
                    "followerCount",
                    "followingCount",
                    "postsCount",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "Update the authenticated user's profile",
        "tags": [
          "User"
        ],
        "parameters": [],
        "operationId": "updateMe",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "maxLength": 50,
                    "nullable": true
                  },
                  "displayName": {
                    "type": "string",
                    "maxLength": 50,
                    "nullable": true
                  },
                  "bio": {
                    "type": "string",
                    "maxLength": 1000,
                    "nullable": true
                  },
                  "avatarKey": {
                    "type": "string",
                    "nullable": true
                  },
                  "bannerKey": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    },
                    "emailId": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "bio": {
                      "type": "string",
                      "nullable": true
                    },
                    "avatarKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "bannerKey": {
                      "type": "string",
                      "nullable": true
                    },
                    "followerCount": {
                      "type": "number"
                    },
                    "followingCount": {
                      "type": "number"
                    },
                    "postsCount": {
                      "type": "number"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "emailId",
                    "displayName",
                    "bio",
                    "avatarKey",
                    "bannerKey",
                    "followerCount",
                    "followingCount",
                    "postsCount",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete the authenticated user's account",
        "tags": [
          "User"
        ],
        "parameters": [],
        "operationId": "deleteMe",
        "responses": {
          "204": {
            "description": "204"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "version": "1.0.0",
    "title": "Nexus REST API - Documentation",
    "description": "Nexus REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}